pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blogDatabaseLib"
version = "1.0.0"
description = "sb"
authors = [
    {name = "KhVeMx", email = "khvemx@foxmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.7"
dependencies = [
    "PyYAML>=6.0.3",
]

[tool.setuptools.packages.find]
where = ["./data/"]
include = ["blogDatabaseLib"]

